# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "db2b-api",
        "description": "Développement d'une plateforme B2B multilingue complète",
        "contact": {
            "email": "ayenaaurel15@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/permission/index": {
            "get": {
                "tags": [
                    "Permission"
                ],
                "summary": "Get all permissions",
                "operationId": "2306a1f461cca99b72b7b4de8fa9ac21",
                "responses": {
                    "200": {
                        "description": "List of permissions"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/permission/store": {
            "post": {
                "tags": [
                    "Permission"
                ],
                "summary": "Create a new permission ",
                "operationId": "aaa9a1489978d8237d358f1010932130",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "create"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Permission  created successfully"
                    },
                    "401": {
                        "description": "Invalid credentials"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/permission/show/{id}": {
            "get": {
                "tags": [
                    "Permission"
                ],
                "summary": "Get a specific permission by ID",
                "operationId": "1cae0ef6ddb328b988765a8129134b9d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the permission",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Permission details"
                    },
                    "404": {
                        "description": "Permission not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/permission/indexbycategorie": {
            "get": {
                "tags": [
                    "Permission"
                ],
                "summary": "Get all permissions groupe by categorie",
                "operationId": "d6606b4331cf95c60b312c1d0977f8fa",
                "responses": {
                    "200": {
                        "description": "List of permissions groupe by category"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/role/index": {
            "get": {
                "tags": [
                    "Role"
                ],
                "summary": "Get all roles",
                "operationId": "b3d674de4d1bfdce32734bddc8628891",
                "responses": {
                    "200": {
                        "description": "List of roles"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/role/store": {
            "post": {
                "tags": [
                    "Role"
                ],
                "summary": "Create a new role ",
                "operationId": "217d536f741ea073dee6d97e5bb36df9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "traveler"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Role  created successfully"
                    },
                    "401": {
                        "description": "Invalid credentials"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/role/show/{id}": {
            "get": {
                "tags": [
                    "Role"
                ],
                "summary": "Get a specific role by ID",
                "operationId": "a40324f1b443d45a408830e2695bbeb9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the role",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Role details"
                    },
                    "404": {
                        "description": "Role not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/role/destroy/{id}": {
            "delete": {
                "tags": [
                    "Role"
                ],
                "summary": "Delete a role by ID",
                "operationId": "fa71958128495cee0ebb0097bad4087f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the role",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Role deleted successfully"
                    },
                    "404": {
                        "description": "Role not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}